Factor out common firearm summary logic to core
